# lesson-booking/openapi.yaml
openapi: 3.1.0
info:
  title: Lesson Booking API
  version: 0.1.0
  description: |
    レッスン予約サービス。複数 Google カレンダー連携で空き枠を出し、40 日先まで予約できる。
    レッスン開始まで 14 日未満のキャンセル・変更はメッセージ+対応方法(3 択)を添えた変更要求となり、主催者の承認で確定する。
    詳細は docs/spec.md。
servers:
  - url: http://localhost:8787
security:
  - bearerAuth: []
  - devHeader: []
...
paths:
...
  /bookings:
    get:
      tags: [student]
      summary: 自分の予約一覧(変更要求履歴付き)。過去分は 60 日前まで、キャンセルフィー未払いは期間に関係なく含む
      parameters: [{ $ref: '#/components/parameters/since' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - { $ref: '#/components/schemas/StudentBooking' }
                    - type: object
                      properties:
                        changeRequests: { type: array, items: { $ref: '#/components/schemas/ChangeRequest' } }
    post:
      tags: [student]
      summary: 予約作成(40 日先まで・空き枠のみ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostId, startAt]
              properties:
                hostId: { type: string }
                startAt: { type: string, format: date-time }
                note: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/StudentBooking' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /bookings/{id}/slots:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get:
      tags: [student]
      summary: 振替先の候補となる空き枠。公開の空き枠と違い、この予約自身(とそのカレンダーの予定)を埋まっている枠に数えない
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots: { type: array, items: { $ref: '#/components/schemas/Slot' } }
                  bookingHorizonDays: { type: integer }
        '403': { $ref: '#/components/responses/Error' }
  /bookings/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get:
      tags: [student]
      summary: 予約詳細(変更要求履歴付き)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/StudentBooking' }
                  - type: object
                    properties:
                      changeRequests: { type: array, items: { $ref: '#/components/schemas/ChangeRequest' } }
  /bookings/{id}/change:
    parameters: [{ $ref: '#/components/parameters/id' }]
    post:
      tags: [student]
      summary: キャンセル・変更
      description: |
        レッスン開始まで 14 日以上: 即時反映(200, type=applied)。
        14 日未満: message と option が必須。変更要求を作成して承認待ち(202, type=pending_approval)。
        kind=reschedule または option=reschedule_within_two_weeks では proposedStartAts(第1〜第3希望、元の日から前後 N 日以内。N は講師の rescheduleRangeDays)が必須。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: { type: string, enum: [cancel, reschedule] }
                message: { type: string, description: 直前変更時は必須 }
                option: { $ref: '#/components/schemas/LateChangeOption' }
                proposedStartAts:
                  type: array
                  maxItems: 3
                  items: { type: string, format: date-time }
                  description: 振替の希望日時(第1希望から順)。直前は 1〜3 件、猶予ありの即時変更は 1 件
                proposedStartAt: { type: string, format: date-time, deprecated: true, description: 旧形式(希望日時 1 件) }
                feeMethod:
                  $ref: '#/components/schemas/FeeMethod'
                  description: option=pay_cancellation_fee のとき必須(kind=cancel のみ)。主催者が受け付けている方法から選ぶ
      responses:
        '200':
          description: 即時反映
          content:
            application/json:
              schema:
                type: object
                properties:
                  type: { type: string, enum: [applied] }
                  booking: { $ref: '#/components/schemas/Booking' }
        '202':
          description: 承認待ち
          content:
            application/json:
              schema:
                type: object
                properties:
                  type: { type: string, enum: [pending_approval] }
                  booking: { $ref: '#/components/schemas/Booking' }
                  request: { $ref: '#/components/schemas/ChangeRequest' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
  /bookings/{id}/fee-checkout:
    parameters: [{ $ref: '#/components/parameters/id' }]
    post:
      tags: [student]
      summary: 未払いキャンセルフィーの決済ページ URL(講師の Stripe アカウントで直接決済)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [successUrl, cancelUrl]
              properties:
                successUrl: { type: string, format: uri }
                cancelUrl: { type: string, format: uri }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { url: { type: string } } } } } }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase Auth の access_token }
    devHeader: { type: apiKey, in: header, name: x-dev-user-email, description: AUTH_MODE=dev のときのみ }
  parameters:
...
    id: { name: id, in: path, required: true, schema: { type: string } }
    since: { name: since, in: query, description: 過去分の起点(既定は 60 日前), schema: { type: string, format: date-time } }
  responses:
    Error:
      description: エラー
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    enum: [validation, outside_booking_window, late_change_requires_request, forbidden, not_found, slot_unavailable, change_request_pending, invalid_state, plan_limit, calendar_error, internal]
                  message: { type: string }
                  details: {}
  schemas:
    FeeMethod:
      type: string
      enum: [card, bank_transfer, in_person]
      description: クレジットカード / 銀行振込 / 次回レッスン時に手渡し
    LateChangeOption:
      type: string
      enum: [request_approval, reschedule_within_two_weeks, pay_cancellation_fee]
      description: 承認を求める / 講師の振替期間内の別日に振替(値の名前は互換のため据え置き)/ キャンセルフィーを支払う
...
    Slot:
      type: object
      properties:
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
...
    Booking:
      type: object
      properties:
        id: { type: string }
        hostId: { type: string }
        studentId: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        status: { type: string, enum: [confirmed, cancelled] }
        calendarEventId: { type: [string, 'null'] }
        note: { type: [string, 'null'] }
        cancellationFeeStatus: { type: string, enum: [none, pending, paid] }
        cancellationFeeAmount: { type: [integer, 'null'] }
        cancellationFeeMethod: { oneOf: [{ $ref: '#/components/schemas/FeeMethod' }, { type: 'null' }] }
        reminderSentAt: { type: [string, 'null'], format: date-time }
        calendarSyncError: { type: [string, 'null'], description: Google カレンダーへの反映に失敗した理由。null なら反映済みか反映不要 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StudentBooking:
      allOf:
        - { $ref: '#/components/schemas/Booking' }
        - type: object
          properties:
            requiresApprovalToChange:
              type: boolean
              description: 今キャンセル・変更すると主催者承認が必要か(開始まで講師の lateChangeThresholdDays 日未満)
            feePayableOnline:
              type: boolean
              description: 未払いのキャンセルフィーをオンラインで払えるか
            bankTransferInfo:
              type: string
              description: 振込で承認された未払いの予約にだけ付く振込先
            rescheduleRangeDays:
              type: integer
              description: この予約の講師の振替期間
            lateChangeThresholdDays:
              type: integer
              description: この予約の講師が、開始の何日前から承認制にしているか
    ChangeRequest:
      type: object
      properties:
        id: { type: string }
        bookingId: { type: string }
        hostId: { type: string }
        studentId: { type: string }
        kind: { type: string, enum: [cancel, reschedule] }
        option: { $ref: '#/components/schemas/LateChangeOption' }
        message: { type: string }
        proposedStartAts: { type: array, items: { type: string, format: date-time }, description: 振替の希望日時(第1希望から順、最大 3) }
        approvedStartAt: { type: [string, 'null'], format: date-time, description: 講師が承認時に選んだ振替先 }
        feeMethod: { oneOf: [{ $ref: '#/components/schemas/FeeMethod' }, { type: 'null' }] }
        status: { type: string, enum: [pending, approved, rejected] }
        decisionNote: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        decidedAt: { type: [string, 'null'], format: date-time }